Refactor part. 2
