+++ assets/js/client/debug_panel/debug_button.js
@@ -33,101 +33,3 @@ export function createDebugButton() {
   tempDiv.innerHTML = debugButtonHtml;
   return tempDiv.firstElementChild;
 }
-
-export function initButtonEvents(debugButton, tooltip, liveDebuggerURL) {
-  // Button state
-  const buttonState = {
-    isDragging: false,
-    tooltipVisible: false,
-  };
-
-  // Button event handlers
-  const onClick = () => {
-    if (buttonState.isDragging) {
-      placeButton();
-    } else {
-      if (buttonState.tooltipVisible) {
-        hideTooltip();
-      } else {
-        showTooltip();
-      }
-    }
-  };
-
-  const startDragging = () => {
-    buttonState.isDragging = true;
-    debugButton.style.cursor = 'grabbing';
-    document.addEventListener('mousemove', onMouseMove);
-  };
-
-  const placeButton = () => {
-    buttonState.isDragging = false;
-    debugButton.style.cursor = 'pointer';
-    document.removeEventListener('mousemove', onMouseMove);
-  };
-
-  const onMouseMove = (event) => {
-    if (buttonState.isDragging) {
-      const buttonWidth = debugButton.offsetWidth;
-      const buttonHeight = debugButton.offsetHeight;
-
-      // Make sure the button doesn't overflow the viewport
-      const maxLeft = window.innerWidth - buttonWidth;
-      const maxTop = window.innerHeight - buttonHeight;
-
-      const newLeft = Math.max(
-        0,
-        Math.min(event.clientX - buttonWidth / 2, maxLeft)
-      );
-      const newTop = Math.max(
-        0,
-        Math.min(event.clientY - buttonHeight / 2, maxTop)
-      );
-
-      debugButton.style.left = `${newLeft}px`;
-      debugButton.style.top = `${newTop}px`;
-      debugButton.style.right = 'auto';
-      debugButton.style.bottom = 'auto';
-    }
-  };
-
-  const showTooltip = () => {
-    const buttonRect = debugButton.getBoundingClientRect();
-    const tooltipWidth = 160;
-    const tooltipHeight = 120;
-
-    // Check if the tooltip would overflow on the right
-    if (buttonRect.right + tooltipWidth > window.innerWidth) {
-      tooltip.style.left = `${buttonRect.left - tooltipWidth}px`;
-    } else {
-      tooltip.style.left = `${buttonRect.right}px`;
-    }
-
-    // Check if the tooltip would overflow on the bottom
-    if (buttonRect.top + tooltipHeight > window.innerHeight) {
-      tooltip.style.top = `${buttonRect.bottom - tooltipHeight}px`;
-    } else {
-      tooltip.style.top = `${buttonRect.top}px`;
-    }
-
-    tooltip.style.display = 'block';
-    buttonState.tooltipVisible = true;
-  };
-
-  const hideTooltip = () => {
-    tooltip.style.display = 'none';
-    buttonState.tooltipVisible = false;
-  };
-
-  // Attach event listeners
-  debugButton.addEventListener('click', onClick);
-
-  // Return state and methods for external access
-  return {
-    ...buttonState,
-    startDragging,
-    placeButton,
-    showTooltip,
-    hideTooltip,
-  };
-}
